requirements/prod.txt
Flask==1.0.2
Flask-GraphQL==2.0.0
PyJWT==1.7.1
Authlib==0.10
gql==0.1.0
graphene==2.1.3
graphql-core==2.1
graphql-relay==0.4.5
graphql-server-core==1.1.1
python-dotenv==0.10.1
six==1.12.0
gunicorn==19.9.0
requests==2.21.0
